Fix parsing for record entries.
Record entries were unable to properly detect duration as a value. The
were viewing it as a number & then a command. In reality the issue was
that the record entry is recursive - so it was thinking of the command
as the start of the new record entry. I fixed that by changing the
record logic altogether by introducing mandatory separators (can be a
space or a newline) between record entries before the final one.